add clustering memory budget with retry batching for threshold optimization
optimize_thresholds() can bound clustering-owned native memory and retry
with fewer threads, more subproblems, then overlap-preserving splits
instead of failing outright. Default behavior is unchanged when no
budget is set.

+++ src/DistclusterWorker.cpp
@@ -11,6 +11,6 @@ size_t DistClusterWorker::aligned() {
   return _aligned;
 }
 
-std::uint8_t DistClusterWorker::n_threads() {
+std::size_t DistClusterWorker::n_threads() {
   return threads;
 }
